pyproject.toml
[project]
name = "sungsimdangbot"
version = "1.6.0"
description = "Telegram bot with various utility features"
requires-python = ">=3.10"
dependencies = [
    "pyTelegramBotAPI>=4.14,<5",
    "requests>=2.28,<3",
    "python-dotenv>=1.0,<2",
    "google-genai>=1.0,<2",
    "pydantic>=2.9,<3",
    "peewee>=3.17,<5",
    "telegramify-markdown>=1.0,<2",
    "tenacity>=8.0,<10",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4",
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pre-commit>=3.0",
]

[tool.setuptools.packages.find]
include = ["config*", "modules*", "resources*"]

[tool.ruff]
line-length = 120
exclude = [".venv", "__pycache__", ".git"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "S113", "FURB", "C4"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.coverage.run]
source = ["modules", "config", "resources"]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
fail_under = 70
